+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- sort labels alphabetically
+
 ## [0.5.1] - 2023-12-05
 
 - fix todo creation overlay not opening